refactor: reorganize project structure by moving core modules into src directory
