+++ src/orchestration/vector_management.py
