+++ src/ui/streamlit.py
@@ -1,4 +1,4 @@
-import streamlit as st
+import src.ui.streamlit as st
 import requests
 import pandas as pd
